# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (one message per field)"
          },
          "409": {
            "description": "Email is already registered"
          }
        },
        "summary": "Create an email + password account; a 6-digit verification code is emailed",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password (same message and timing for every failure)"
          }
        },
        "summary": "Log in with email + password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          },
          "401": {
            "description": "Unknown, expired, revoked or reused refresh token"
          }
        },
        "summary": "Swap a refresh token for a new pair (rotation). Reusing an old one revokes every session",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "End this session (always 204)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Verified (also if already verified)"
          },
          "400": {
            "description": "Invalid code, expired code, or too many attempts"
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify your email with the 6-digit code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/resend-verification": {
      "post": {
        "operationId": "AuthController_resendVerification",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Email is already verified"
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "429": {
            "description": "Wait before requesting another code (60 s)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Email a new verification code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Email a password-reset code if the account exists (always 204: never reveals whether it does)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Invalid or expired code (one message for every failure)"
          }
        },
        "summary": "Set a new password with the emailed code; ends every session",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/oauth/exchange": {
      "post": {
        "operationId": "AuthController_oauthExchange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthExchangeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid, expired or already used code, or wrong verifier"
          }
        },
        "summary": "End of a social login: trade the single-use code from the return link + your PKCE verifier for tokens",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/google/start": {
      "get": {
        "operationId": "SocialAuthController_googleStart",
        "parameters": [
          {
            "name": "redirect",
            "required": true,
            "in": "query",
            "description": "App deep link to return to (exp://… or musicroom://…)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code_challenge",
            "required": true,
            "in": "query",
            "description": "PKCE S256 challenge: base64url(SHA-256(verifier)), 43 chars",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to Google"
          },
          "400": {
            "description": "redirect is not an app deep link, or bad code_challenge"
          }
        },
        "summary": "Open in a browser tab: redirects to Google sign-in (OpenID Connect)",
        "tags": [
          "Social login (browser redirects)"
        ]
      }
    },
    "/auth/google/callback": {
      "get": {
        "operationId": "SocialAuthController_googleCallback",
        "parameters": [
          {
            "name": "code",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirects the browser back to the app with ?code= (single-use, 60 s) or ?error=account_exists|email_required|access_denied|…"
          },
          "400": {
            "description": "Forged or expired state"
          }
        },
        "summary": "Google sends the browser here (not for direct calls)",
        "tags": [
          "Social login (browser redirects)"
        ]
      }
    },
    "/auth/facebook/start": {
      "get": {
        "operationId": "SocialAuthController_facebookStart",
        "parameters": [
          {
            "name": "redirect",
            "required": true,
            "in": "query",
            "description": "App deep link to return to (exp://… or musicroom://…)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code_challenge",
            "required": true,
            "in": "query",
            "description": "PKCE S256 challenge: base64url(SHA-256(verifier)), 43 chars",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to Facebook"
          },
          "400": {
            "description": "redirect is not an app deep link, or bad code_challenge"
          }
        },
        "summary": "Open in a browser tab: redirects to Facebook login (OAuth 2.0)",
        "tags": [
          "Social login (browser redirects)"
        ]
      }
    },
    "/auth/facebook/callback": {
      "get": {
        "operationId": "SocialAuthController_facebookCallback",
        "parameters": [
          {
            "name": "code",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirects the browser back to the app with ?code= (single-use, 60 s) or ?error=account_exists|email_required|access_denied|…"
          },
          "400": {
            "description": "Forged or expired state"
          }
        },
        "summary": "Facebook sends the browser here (not for direct calls)",
        "tags": [
          "Social login (browser redirects)"
        ]
      }
    },
...
  },
  "info": {
    "title": "Music Room API",
    "description": "REST API of Music Room (Track Vote events, accounts, music search).\n\n**Auth:** log in with `POST /auth/login` (or a social login), then click **Authorize** and paste the `accessToken`.\nAccess tokens last 15 minutes; get a new pair with `POST /auth/refresh`.\n\n**Real-time** (Socket.IO namespace `/events`): see `docs/realtime.md` in the repository.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "The accessToken from /auth/login"
      }
    },
    "schemas": {
      "CurrentUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "baaa7e88-cac5-41d3-b29d-4b3e64941eec"
          },
          "email": {
            "type": "string",
            "description": "Always lower-case.",
            "example": "alice@example.com"
          },
          "displayName": {
            "type": "string",
            "example": "Alice"
          },
          "emailVerified": {
            "type": "boolean",
            "description": "False until the user enters the 6-digit code emailed at registration."
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "emailVerified"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72
          },
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "displayName"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "TokenPairDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Opaque, single-use refresh token (7 days). Rotated on every POST /auth/refresh.",
            "example": "Qx3…"
          },
          "accessToken": {
            "type": "string",
            "description": "Short-lived JWT (15 min). Send as `Authorization: Bearer <accessToken>`."
          }
        },
        "required": [
          "refreshToken",
          "accessToken"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$"
          }
        },
        "required": [
          "code"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72
          }
        },
        "required": [
          "email",
          "code",
          "newPassword"
        ]
      },
      "OAuthExchangeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "codeVerifier": {
            "type": "string",
            "pattern": "^[A-Za-z0-9\\-._~]{43,128}$"
          }
        },
        "required": [
          "code",
          "codeVerifier"
        ]
      },
...
    }
  }
}